Add notes for watching progress
Realistically this is very, very slow. Progress tracking through alphabet is helpful.

Option for speeding up is to add optimization to grow out from a collapse point to identify more indices for deletion. To be cautious this could be limited to when there is a single pair for deletion. This would have been particularly useful in part 2 where there were dozens (possible over 100) of string updates that were the removal of nested pairs. Each output string creation in this approach is very expensive.